python-api/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
stripe==5.4.0
google-cloud-storage==2.9.0
onnxruntime==1.14.1
torch==2.0.1
transformers==4.29.2
diffusers==0.16.1
accelerate==0.19.0
safetensors==0.3.1
python-dotenv==1.0.0
sqlalchemy==2.0.15
alembic==1.11.1
psycopg2-binary==2.9.6
aiofiles==23.1.0
httpx==0.24.1
pytest==7.3.1
pytest-asyncio==0.21.0
tenacity==8.2.2
ratelimit==2.2.1
